Budget - rounding numbers

+++ eu2020/common/budget.py
@@ -56,10 +56,10 @@ class Budget:
     def update_balance(self) -> None:
         self.budget["dept"] -= self.get_income() + self.budget["extra_income"] \
             - self.get_outcome() - self.budget["extra_outcome"] \
-            - self.get_guarantee() * 0.25
+            - int(self.get_guarantee() * 0.25)
         self.budget["extra_income"] = 0
         self.budget["extra_outcome"] = 0
-        self.budget["guarantee"] = self.get_guarantee() * 0.5
+        self.budget["guarantee"] = int(self.get_guarantee() * 0.5)
 
     def get_budget_report(self):
         result = ""
